Q(Styled)ItemDelegate: disable style animations
The style animation system expects a single QObject per animated item.
Therefore, this change disables style animations for checkboxes etc.
within itemviews, because the view cannot act as an animation target
for several checkboxes. The result of that kind of animation would be
randomly checked/unchecked checkboxes, or sometimes even left in the
middle of the transition.

Task-number: QTBUG-30146
Change-Id: I1d7165571c723e84a0e6fbda266cb03b04deabd8
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@digia.com>
Reviewed-by: Jens Bache-Wiig <jens.bache-wiig@digia.com>