Add support for the Clang compiler bundled with XCode
Since c45595d64831990311f92fcebc4e34e2797f5352, we use the toolchain
based on the XCode SDK, not the one in $PATH. Turns out that the Clang
that comes bundled with XCode 4.6 reports its version as

     Apple LLVM version 4.2 (clang-425.0.27) (based on LLVM 3.2svn)

Instead of "Apple clang". So we need to match for (clang|LLVM).

Extended regular expressions (with -E) were necessary because the sed
that comes with Mac OS X is apparently broken and will not work with
\(clang\|LLVM\). GNU sed accepts -E as an alias for -r, meaning
extended regexps.

Change-Id: I5a15de30721216b086c3d39a080cc6496c503985
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@digia.com>