Windows: Fix the last file dialog bottleneck.
Went from taking 30 seconds to 2 seconds, on a SDCard with 10k files.

Windows file dialog does not resolve NTFS symlinks,
it just shows an empty icon, and the link name, not the target.

This allows for a big performance gain by reducing the number of
calls to GetFileAttributesEx() by checking the extension directly.

This also fixes the problems with the native file dialog, which
for some reason, is creating a QFileSystemModel too.

Task-number: QTBUG-13182
Change-Id: Ie2739765fd6c7daea64e3cf1d208ba9720bd39f2
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@digia.com>
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>