Android: Backslash workarounds for Unix Makefile generator
If mingw32-make.exe encounters a backslash as the last character
on a line it interprets this as signifying line continuation.

When building Android Qt on Windows via cmd.exe, this happens
as backslashes are used on Windows as directory separators.

The workarounds are to make sure that a comment appears directly
after the definition of DESTDIR and that a space ends such
$(MOVE) command lines.

Change-Id: I7f93b655e004edaadac41d0d96bca23e1ba3a85c
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>