Android: handle keyPress event for Key_Back
Added logic so that accepting either the press or the release will
keep the app running. This makes it possible to use the onBackPressed
functionality in QML.

This functionality is only intended for running in a complete Android
environment, so make sure that we don't terminate the application in
the NO_SDK case.

Task-number: QTBUG-30803

Change-Id: I2546eea73bf6a6ee8b196125b7556479b9b10a9c
Reviewed-by: BogDan Vatra <bogdan@kde.org>